cpi_targets_ml/requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn==1.0.2

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
fastapi==0.101.0

# utilities
six>=1.14
joblib
memoized-property
termcolor

# web server
uvicorn==0.23.2


jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-resource-usage==0.6.1
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.



pydata-google-auth==1.3.0
googletrans==3.1.0a0
googleapis-common-protos==1.56.4
google-trans-new==1.1.9
google-resumable-media==2.4.0
google-pasta==0.2.0
google-crc32c==1.3.0
google-cloud==0.34.0
google-cloud-vision==3.4.1
google-cloud-secret-manager==2.12.6
google-cloud-dataproc==5.0.3
google-cloud-core==2.3.2
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.10.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.0.3
google-api-python-client==1.7.9
google-api-core==2.11.0
googletrans==3.1.0a0
webcolors==1.13
oauth2client==4.1.2
nltk==3.6.7
scikit-learn==1.0.2
pyarrow==6.0.1
tqdm==4.62.3
colorama
